requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
google-generativeai==0.8.0
python-dotenv==1.0.1
